2021.8.22

+++ server/model/z_outcome.go
@@ -6,7 +6,7 @@ type Outcome struct {
 	gorm.Model
 	BillID				uint			`json:"billId" form:"billId"`
 
-	Money				float32			`json:"money" form:"money"`
+	Money				float64			`json:"money" form:"money"`
 	Comment 			string			`json:"comment" form:"comment" gorm:"type:varchar(512)"`
 }
